refactored server into own class
